Update jquery-ui from 1.10 to 1.11
They no longer seem to be uploading to NPM, so i'm using the github
repo as the source. They do upload versions to bower, but that wouldn't
work for us, as this tagging package is pulled into the main UI via
npm, which of course wouldn't resolve bower deps.

I've also changed the require to include just the dialog functionality,
rather than everything, which should cut down the built size somewhat.
The previous version was taking ~500KiB IIRC.